Compatable libraries

+++ src/kernel/mem/heap.c
@@ -1,5 +1,4 @@
 #include "heap.h"
-#include "../std/string.h"
 
 static u64 data_ptr = HEAP_MEMORY_START_POINT + MEMTABLE_SIZE;
 static u64 heap_ptr = HEAP_MEMORY_START_POINT;
@@ -19,7 +18,7 @@ void* malloc(size_t size) {
     h.endPoint = data_ptr + size;
     data_ptr += size;
 
-    memcpy((void*) heap_ptr, (void*) &h, sizeof(Heap));
+    memcpy((void*) (u32) heap_ptr, (void*) &h, sizeof(Heap));
     heap_ptr += sizeof(Heap);
 
     void* ret_ptr = (void*) h.startPoint;
@@ -30,7 +29,7 @@ void free(void *restrict ptr) {
     u64 startPtr = HEAP_MEMORY_START_POINT;
     Heap h;
     for(u32 i = 0; i < HEAP_COUNT; i++) {
-        memcpy(&h, (void*)startPtr, sizeof(Heap));
+        memcpy(&h, (void*) (u32) startPtr, sizeof(Heap));
         startPtr += sizeof(Heap);
         if(((void*) h.startPoint) == ptr) {
             h.allocated = false;
